Se separan funcionalidades entre ItemList e ItemListContainer
